refactor(engine): extract Sampling helpers + Stop controller — final split
Moves 6 decode-loop tail methods to src/runtime/engine_sampling_stop.cpp:
  - is_stop_token, track_think_state, should_stop (stop check)
  - fill_sampling_params, upload_penalties, fill_recurrent_state (per-request setup)

Two related clusters colocated because they share the per-request
state-passing pattern (Request& + InferenceState&) and run in the
decode loop tail.

This is the FINAL per-subsystem extraction of Phase 4. After this PR,
engine.cpp is at ~570 LOC — well under the <=800 LOC plan target.

Declarations in engine.h unchanged. Bodies byte-identical.

Phase 4 of docs/superpowers/specs/2026-05-20-architecture-refactor-roadmap-design.md
